requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.6.0
django-crum==0.7.9
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
django-rest-passwordreset==1.5.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.49
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
